+++ CHANGELOG.md
@@ -34,6 +34,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Fixed a small bug which would cause Group Rolls not to behave properly
 
+### Removed
+
+- Removed Handlebars helpers that overwrote helpers defined by Foundry core
+
 ## [v0.13]
 
 ### Added